Math-Study-Assistant/requirements.txt
pypdf 
pinecone 
tiktoken 
orjson 
python-dotenv 
google.generativeai